Add insert and fetch for database
Use "virtual pages" i.e. keeped them in memory only, to save rows to the
database. This also creates a new page per insert (should change). It
also has a static schema for just int values (milestone 1). And it does
a linear search through the page (of one value), which is linear because
it's constant, but needs to change after we add many rows to a page.