record lock approach
prob has issues stemming from the fact that you can't really hold a lock on something that's been deleted (we try to release lock, but don't validate that it was in the case that it was deleted), or hasn't been inserted yet (we obtain the lock right after inserting).

also update rollbacks actually insert another version of the record, instead of readjusting the indirection on the base back one.

testAbort.py best shows that this works, updates are rolled back since an insert fails (duplicate primary key)